Improve tumblebug client
